% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ollama}
\alias{check_ollama}
\title{Check Ollama Installation}
\usage{
check_ollama(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Should informative messages be printed to the console? Default is TRUE.}
}
\value{
Logical TRUE if Ollama is installed and responding; otherwise returns FALSE. Also prints diagnostic messages if verbose = TRUE.
}
\description{
Verify whether the Ollama backend is properly installed and available by testing the connection and retrieving the list of local models.
}
\details{
This function internally calls \code{ollamar::test_connection()} and \code{ollamar::list_models()} to check the availability of the Ollama service and confirm that at least one local model is installed.
}
\examples{
\dontrun{
check_ollama()
}

}
\keyword{ollama}
